Hub interactions- talk (#650)
* Add hub sublocation tutorial and input routing

Introduce HubSublocationTutorialHandler and a HubSublocation prefab to display ordered UI tutorial pages on first visit and hand input back when complete. Route input to the active tutorial by adding ActiveTutorialHandler to SpecificUiHandler and short-circuiting input in HubManager and SpecificUiHandler while a tutorial is active. HubSubLocation now tracks tutorial state and disables accepting input until the tutorial finishes. Also add GamewideContextBrain.GetOrCreateAvatarInstance and update HubCharacterManager to build the avatar model from the persistent roster instance, with warning logs when roster/avatar/model are missing.

* fix avatar bug

* Replenish item uses after battles

Add post-battle item replenishment and related fixes across systems. InventoryBrain: subscribe to OnBattleCompleted and restore item uses based on new ReplenishUseType logic. ObjectItem: expose ReplenishUsesAfterBattle and ReplenishUsesAfterBattleAmount properties. ObjectItemInstance: add ReplenishUses(amount) to restore item uses. Combat AI: BattleContextAIHelper listens for last-attacker events and exposes retaliation state; GoalEvaluation boosts aggression and emotion for retaliatory behavior. Hub UI: hide POIs when returning to hub and skip fade when POI is inactive. ScreenKeyboard: remove unused buttonSpacing field. RadialMenu: use inputRepeatDelay for navigationRepeatDelay. SceneSkyboxSetter: remove an unused field and throttle night-tint updates to avoid frequent updates.

* hub character interaction menu

Implements a full character actions menu and input handling: adds CharacterInteractionOption/Type, navigation, selection, availability checks (train/talk/meal/spa/dance/gift/lost/support/recruit) and wiring to HubCharacterManager. For integration, HubCharacterInteraction now validates refs and exposes Show/Hide menu and HandleInput; SpecificUiHandler forwards input to the interaction component. Adds Spa POI type and a minimal Spa MonoBehaviour. Smaller changes: make CharacterData.IsAllyOrRecruitable public and include ALLY in recruitable UI logic; create a fresh runtime player roster instance in GamewideContext persistence to avoid misleading null logs; add ChitChat one-shot type; remove some LogInfo calls and an unused local in dock code; expose _brain on HubCharacterManager (HideInInspector); remove unused using in ScreenKeyboard. Overall these changes wire up the hub character action UX and clean up related plumbing.

* fix input duplication

* working on support

* Update Demos

* clean  up + interaction events

* persist more to hub day state

+++ Assets/TurnrootFramework/Characters/Components/InstanceComponents/CharacterSupportSystem.cs
@@ -46,7 +46,7 @@ namespace Turnroot.Characters
             return OperationResult.Successful();
         }
 
-        internal OperationResult IncreaseSupport(CharacterData character, int amount)
+        internal OperationResult IncreaseSupport(CharacterData character, float amount)
         {
             bool ok = ValidationHelper.ValidateNotNull(
                 "CharacterInstance.IncreaseSupport",
